Extract Css import rewriter
